Add GetSiteNumIfAny(Direct|Sym)
This allows any (ulam) Atom (represented by its BitStorage<EC>
to determine which eventwindow -- if any -- it is stored at.

This is a perhaps questionable functional extension, since
implementing it might not be trivial depending on internals,
but it's been wanted for just too long.  As we increasingly
dish to methods on atoms other that ew[0], those methods need
to be able to find out where they are and navigate to their
neighbors, if need be.